Refactoring: move DNS resolving into a separate file.
