+++ src/main/java/org/masukomi/aspirin/core/RemoteDelivery.java
@@ -68,11 +68,9 @@ import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.Properties;
-import java.util.Vector;
 
 import javax.mail.MessagingException;
 import javax.mail.Session;
@@ -86,11 +84,7 @@ import org.apache.commons.pool.ObjectPool;
 import org.apache.james.core.MailImpl;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailAddress;
-import org.xbill.DNS.Lookup;
-import org.xbill.DNS.MXRecord;
-import org.xbill.DNS.Record;
-import org.xbill.DNS.TextParseException;
-import org.xbill.DNS.Type;
+import org.masukomi.aspirin.core.dns.DnsResolver;
 
 /**
  * <p>This thread </p>
@@ -114,7 +108,7 @@ public class RemoteDelivery extends Thread implements ConfigurationChangeListene
 	protected QuedItem qi;
 	protected MailQue que;
 	
-	private static final String SMTPScheme = "smtp://";
+//	private static final String SMTPScheme = "smtp://";
 	
 	// TODO This is a temporary constructor, it should be changed
 	public RemoteDelivery(ThreadGroup parentThreadGroup) {
@@ -210,7 +204,7 @@ public class RemoteDelivery extends Thread implements ConfigurationChangeListene
 			try {
 				// targetServers = MXLookup.urlsForHost(host); // farking
 				// unreliable jndi bs
-				targetServers = getMXRecordsForHost(host);
+				targetServers = DnsResolver.getMXRecordsForHost(host);
 			} catch (Exception e) {
 				log.error(getClass().getSimpleName()+" ("+getName()+" ).deliver(): Could not get MX for "+host+".",e);
 			}
@@ -709,111 +703,111 @@ public class RemoteDelivery extends Thread implements ConfigurationChangeListene
 		this.que = que;
 	}
 
-	/**
-	 * <p>This method gives back the host name(s) where we can send the email.
-	 * </p>
-	 * 
-	 * <p>First time we ask DNS to find MX record(s) of a domain name. If no MX 
-	 * records are found, we check the upper level domains (if exists). At last 
-	 * we try to get the domain A record, because the MX server could be same as 
-	 * the normal domain handler server. If only upper level domain has MX 
-	 * record then we append the A record of original hostname (if exists) as 
-	 * first element of record collection. If none of these tries are 
-	 * successful, we give back an empty collection.</p>
-	 * 
-	 * Special Thanks to Tim Motika (tmotika at ionami dot com) for 
-	 * his reworking of this method.
-	 * 
-	 * @param hostName We search the associated MX server of this hostname.
-	 * @return Collection of URLName objects. If no MX server found, then it 
-	 * gives back an empty collection.
-	 * 
-	 * TODO public -> private
-	 * 
-	 */
-	public Collection<URLName> getMXRecordsForHost(String hostName) {
-
-		Vector<URLName> recordsColl = null;
-		try {
-			boolean foundOriginalMX = true;
-			Record[] records = new Lookup(hostName, Type.MX).run();
-			
-			/*
-			 * Sometimes we should send an email to a subdomain which does not 
-			 * have own MX record and MX server. At this point we should find an 
-			 * upper level domain and server where we can deliver our email.
-			 *  
-			 * Example: subA.subB.domain.name has not own MX record and 
-			 * subB.domain.name is the mail exchange master of the subA domain 
-			 * too.
-			 */
-			if( records == null || records.length == 0 )
-			{
-				foundOriginalMX = false;
-				String upperLevelHostName = hostName;
-				while(		records == null &&
-							upperLevelHostName.indexOf(".") != upperLevelHostName.lastIndexOf(".") &&
-							upperLevelHostName.lastIndexOf(".") != -1
-					)
-				{
-					upperLevelHostName = upperLevelHostName.substring(upperLevelHostName.indexOf(".")+1);
-					records = new Lookup(upperLevelHostName, Type.MX).run();
-				}
-			}
-
-            if( records != null )
-            {
-            	// Sort in MX priority (higher number is lower priority)
-                Arrays.sort(records, new Comparator<Record>() {
-                    @Override
-                    public int compare(Record arg0, Record arg1) {
-                        return ((MXRecord)arg0).getPriority()-((MXRecord)arg1).getPriority();
-                    }
-                });
-                // Create records collection
-                recordsColl = new Vector<URLName>(records.length);
-                for (int i = 0; i < records.length; i++)
-				{ 
-					MXRecord mx = (MXRecord) records[i];
-					String targetString = mx.getTarget().toString();
-					URLName uName = new URLName(
-							RemoteDelivery.SMTPScheme +
-							targetString.substring(0, targetString.length() - 1)
-					);
-					recordsColl.add(uName);
-				}
-            }else
-            {
-            	foundOriginalMX = false;
-            	recordsColl = new Vector<URLName>();
-            }
-            
-            /*
-             * If we found no MX record for the original hostname (the upper 
-             * level domains does not matter), then we add the original domain 
-             * name (identified with an A record) to the record collection, 
-             * because the mail exchange server could be the main server too.
-			 * 
-			 * We append the A record to the first place of the record 
-			 * collection, because the standard says if no MX record found then 
-			 * we should to try send email to the server identified by the A 
-			 * record.
-             */
-			if( !foundOriginalMX )
-			{
-				Record[] recordsTypeA = new Lookup(hostName, Type.A).run();
-				if (recordsTypeA != null && recordsTypeA.length > 0)
-				{
-					recordsColl.add(0, new URLName(RemoteDelivery.SMTPScheme + hostName));
-				}
-			}
-
-		} catch (TextParseException e) {
-			log.warn(getClass().getSimpleName()+" ("+getName()+").getMXRecordsForHost(): Failed get MX record.",e);
-		}
-
-		return recordsColl;
-	}
+//	/**
+//	 * <p>This method gives back the host name(s) where we can send the email.
+//	 * </p>
+//	 * 
+//	 * <p>First time we ask DNS to find MX record(s) of a domain name. If no MX 
+//	 * records are found, we check the upper level domains (if exists). At last 
+//	 * we try to get the domain A record, because the MX server could be same as 
+//	 * the normal domain handler server. If only upper level domain has MX 
+//	 * record then we append the A record of original hostname (if exists) as 
+//	 * first element of record collection. If none of these tries are 
+//	 * successful, we give back an empty collection.</p>
+//	 * 
+//	 * Special Thanks to Tim Motika (tmotika at ionami dot com) for 
+//	 * his reworking of this method.
+//	 * 
+//	 * @param hostName We search the associated MX server of this hostname.
+//	 * @return Collection of URLName objects. If no MX server found, then it 
+//	 * gives back an empty collection.
+//	 * 
+//	 * TO-DO public -> private
+//	 * 
+//	 */
+//	public Collection<URLName> getMXRecordsForHost(String hostName) {
+//
+//		Vector<URLName> recordsColl = null;
+//		try {
+//			boolean foundOriginalMX = true;
+//			Record[] records = new Lookup(hostName, Type.MX).run();
+//			
+//			/*
+//			 * Sometimes we should send an email to a subdomain which does not 
+//			 * have own MX record and MX server. At this point we should find an 
+//			 * upper level domain and server where we can deliver our email.
+//			 *  
+//			 * Example: subA.subB.domain.name has not own MX record and 
+//			 * subB.domain.name is the mail exchange master of the subA domain 
+//			 * too.
+//			 */
+//			if( records == null || records.length == 0 )
+//			{
+//				foundOriginalMX = false;
+//				String upperLevelHostName = hostName;
+//				while(		records == null &&
+//							upperLevelHostName.indexOf(".") != upperLevelHostName.lastIndexOf(".") &&
+//							upperLevelHostName.lastIndexOf(".") != -1
+//					)
+//				{
+//					upperLevelHostName = upperLevelHostName.substring(upperLevelHostName.indexOf(".")+1);
+//					records = new Lookup(upperLevelHostName, Type.MX).run();
+//				}
+//			}
+//
+//            if( records != null )
+//            {
+//            	// Sort in MX priority (higher number is lower priority)
+//                Arrays.sort(records, new Comparator<Record>() {
+//                    @Override
+//                    public int compare(Record arg0, Record arg1) {
+//                        return ((MXRecord)arg0).getPriority()-((MXRecord)arg1).getPriority();
+//                    }
+//                });
+//                // Create records collection
+//                recordsColl = new Vector<URLName>(records.length);
+//                for (int i = 0; i < records.length; i++)
+//				{ 
+//					MXRecord mx = (MXRecord) records[i];
+//					String targetString = mx.getTarget().toString();
+//					URLName uName = new URLName(
+//							RemoteDelivery.SMTPScheme +
+//							targetString.substring(0, targetString.length() - 1)
+//					);
+//					recordsColl.add(uName);
+//				}
+//            }else
+//            {
+//            	foundOriginalMX = false;
+//            	recordsColl = new Vector<URLName>();
+//            }
+//            
+//            /*
+//             * If we found no MX record for the original hostname (the upper 
+//             * level domains does not matter), then we add the original domain 
+//             * name (identified with an A record) to the record collection, 
+//             * because the mail exchange server could be the main server too.
+//			 * 
+//			 * We append the A record to the first place of the record 
+//			 * collection, because the standard says if no MX record found then 
+//			 * we should to try send email to the server identified by the A 
+//			 * record.
+//             */
+//			if( !foundOriginalMX )
+//			{
+//				Record[] recordsTypeA = new Lookup(hostName, Type.A).run();
+//				if (recordsTypeA != null && recordsTypeA.length > 0)
+//				{
+//					recordsColl.add(0, new URLName(RemoteDelivery.SMTPScheme + hostName));
+//				}
+//			}
+//
+//		} catch (TextParseException e) {
+//			log.warn(getClass().getSimpleName()+" ("+getName()+").getMXRecordsForHost(): Failed get MX record.",e);
+//		}
+//
+//		return recordsColl;
+//	}
 	
 	public void shutdown() {
 		log.trace(getClass().getSimpleName()+" ("+getName()+").shutdown(): Called.");
