+++ src/test/java/org/masukomi/aspirin/core/DNSJavaTest.java
@@ -10,10 +10,13 @@ import java.util.Collection;
 import javax.mail.URLName;
 
 import junit.framework.TestCase;
+
+import org.masukomi.aspirin.core.dns.DnsResolver;
 /**
  * <p>Test of DNS resolving.</p>
  *
  * @author masukomi (masukomi at masukomi dot org)
+ * @deprecated
  *
  */
 public class DNSJavaTest extends TestCase {
@@ -26,20 +29,18 @@ public class DNSJavaTest extends TestCase {
 	public void testDNSLookup() {
 		System.out.println("three problematic domains for MX Record retreival");
 
-		final RemoteDelivery rd = new RemoteDelivery(new ThreadGroup("tempThreadGroup"));
-
 		System.out.println("testing gmx.net");
-		final Collection<URLName> mxRecords1 = rd.getMXRecordsForHost("gmx.net");
+		final Collection<URLName> mxRecords1 = DnsResolver.getMXRecordsForHost("gmx.net");
 		assertNotNull(mxRecords1);
 		assertTrue(mxRecords1.size() > 0);
 
 		System.out.println("testing green.ch");
-		final Collection<URLName> mxRecords2 = rd.getMXRecordsForHost("green.ch");
+		final Collection<URLName> mxRecords2 = DnsResolver.getMXRecordsForHost("green.ch");
 		assertNotNull(mxRecords2);
 		assertTrue(mxRecords2.size() > 0);
 
 		System.out.println("testing tschannen.cc");
-		final Collection<URLName> mxRecords3 = rd.getMXRecordsForHost("tschannen.cc");
+		final Collection<URLName> mxRecords3 = DnsResolver.getMXRecordsForHost("tschannen.cc");
 		assertNotNull(mxRecords3);
 		assertTrue(mxRecords3.size() > 0);
 	}
